Remove esutil package. Move some of the craft that is still used for bootstrapping to esboot package, since other parts do not use it and should not.
