+++ internal/pkg/esboot/esutil.go
@@ -2,7 +2,7 @@
 // or more contributor license agreements. Licensed under the Elastic License;
 // you may not use this file except in compliance with the Elastic License.
 
-package esutil
+package esboot
 
 import (
 	"encoding/json"
